Mock up single toolbar over editor
